requirements.txt
absl-py==1.2.0
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
cachetools==5.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.1.0
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
executing==0.10.0
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
fonttools==4.36.0
future==0.18.2
glob2==0.7
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==2.10
imageio==2.21.1
imageio-ffmpeg==0.4.7
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
keyboard==0.13.5
kiwisolver==1.3.1
Markdown==3.4.1
MarkupSafe==2.1.1
matlab==0.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
moviepy==1.0.3
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
proglog==0.1.10
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pymatlab==0.2.3
pynput==1.7.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
roboflow==0.2.14
rsa==4.9
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
stack-data==0.4.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.6
utils==1.0.1
wcwidth==0.2.5
Werkzeug==2.2.2
wget==3.2
yarg==0.1.9
